+++ Changelog.md
@@ -8,6 +8,7 @@
 * Added alert and icon blade components
 * Removed unnecesary coupon type details view
 * Only show webling booking button if configured
+* Removed google books PHP library, replaced with Javascript solution
 
 ## 2.29.0